fix+perf: ATLAS_SSM_MULTI_SEQ_BATCHED — full per-seq → batched + FP32 fix
Two changes bundled because (1) is a prerequisite for (2):

1. **Correctness fix in the existing multi-seq kernel path (b9e7c28).**
   `gated_delta_rule_decode_f32_multi_seq` in common/ declared BF16
   query/key/value but conv1d_update_l2norm_f32_multi_seq writes FP32.
   Single-seq decode worked because qwen3.6-27b ships a model-specific
   `gated_delta_rule_decode_f32` override with FP32 inputs — I missed
   that pattern when writing the multi-seq variant. Result: gate=1 ON
   produced garbage `!!!` tokens at N≥2 concurrent decode. Fixed by
   switching the multi-seq kernel inputs to FP32 (matches single-seq
   override semantics) and changing the q/k/v strides from BF16 to FP32
   elements at the call site.

2. **Batch the remaining per-seq launches in the multi-seq path.**
   On top of the (now-correct) conv1d + gdn multi-seq launches:
   - **QKVZ**: per-seq w4a16_gemv → single w4a16_gemm at M=num_seqs
     (sequential_qkvz + nvfp4 path; per-seq fallback for other layouts).
   - **BA + GDN gates**: per-seq fused dense_gemv_ba_gates → ONE
     dense_gemm (BA, M=num_seqs, BF16 scratch in ssm_conv_out_f32 head)
     + ONE compute_gdn_gates_multi_seq.
   - **Gated RMS norm**: per-seq gated_rms_norm → ONE
     gated_rms_norm_f32_multi_seq (new kernel — per-head norm at
     grid=(num_v_heads, num_seqs), with per-seq input/gate/output
     strides). Writes value_dim-contig output for batched out_proj.
   - **out_proj**: per-seq w4a16_gemv → single w4a16_gemm at M=num_seqs
     reading the value_dim-contig output from batched RMS.

   Saves 4*(N-1) - 1 launches per SSM layer per token at num_seqs=N.
   At N=4 × 48 SSM layers ≈ 528 launches/token removed.

Correctness validated: N=1 and N=3 concurrent decode (counts 1-15 /
11-25 / 21-35) produce perfect output. Bench at the chosen prompt/CONC
mix shows N=1..6 throughput within ±1% of gate=OFF baseline — the
launch savings are real (visible under cuda profiler) but masked by
the K=3 verify overhead at single-stream concurrent decode. The
infrastructure unlock + correctness fix are the real wins.

Kernels:
- kernels/gb10/common/gated_delta_rule_f32_multi_seq.cu (FP32 inputs)
- kernels/gb10/common/gated_rms_norm_f32_multi_seq.cu (new, per-head)

+++ kernels/gb10/common/gated_delta_rule_f32_multi_seq.cu
@@ -24,11 +24,14 @@
 // ============================================================
 // DECODE multi-seq: gated delta rule recurrent update (FP32 output).
 // ============================================================
+// Inputs are FP32 (NOT BF16): conv1d_update_l2norm_f32_multi_seq writes
+// FP32, and AEON-Q36-27B's calibration relies on FP32 conv→GDN precision.
+// Strides for q/k/v are in FP32 elements (qkvz_size FP32 per seq).
 extern "C" __global__ void gated_delta_rule_decode_f32_multi_seq(
     float* const* __restrict__ h_states,
-    const __nv_bfloat16* __restrict__ query,
-    const __nv_bfloat16* __restrict__ key,
-    const __nv_bfloat16* __restrict__ value,
+    const float* __restrict__ query,
+    const float* __restrict__ key,
+    const float* __restrict__ value,
     const float* __restrict__ gate,
     const float* __restrict__ beta,
     float* __restrict__ output,
@@ -38,8 +41,8 @@ extern "C" __global__ void gated_delta_rule_decode_f32_multi_seq(
     unsigned int k_dim,
     unsigned int v_dim,
     unsigned int gate_beta_stride, // FP32 elements between seqs in gate/beta
-    unsigned int qk_stride,        // BF16 elements between seqs in query/key
-    unsigned int v_in_stride,      // BF16 elements between seqs in value
+    unsigned int qk_stride,        // FP32 elements between seqs in query/key
+    unsigned int v_in_stride,      // FP32 elements between seqs in value
     unsigned int v_out_stride      // FP32 elements between seqs in output
 ) {
     const unsigned int vh  = blockIdx.x;
@@ -53,9 +56,9 @@ extern "C" __global__ void gated_delta_rule_decode_f32_multi_seq(
 
     float* H = h_states[seq] + (vh * k_dim * v_dim);
 
-    const __nv_bfloat16* q_ptr = query + (unsigned long long)seq * qk_stride + kh * k_dim;
-    const __nv_bfloat16* k_ptr = key   + (unsigned long long)seq * qk_stride + kh * k_dim;
-    const __nv_bfloat16* v_ptr = value + (unsigned long long)seq * v_in_stride + vh * v_dim;
+    const float* q_ptr = query + (unsigned long long)seq * qk_stride + kh * k_dim;
+    const float* k_ptr = key   + (unsigned long long)seq * qk_stride + kh * k_dim;
+    const float* v_ptr = value + (unsigned long long)seq * v_in_stride + vh * v_dim;
 
     float g_raw = gate[seq * gate_beta_stride + vh];
     const float g = fminf(fmaxf(g_raw, 0.0f), 1.0f);
@@ -65,13 +68,13 @@ extern "C" __global__ void gated_delta_rule_decode_f32_multi_seq(
     __shared__ float smem_q[128];
 
     if (tid < k_dim) {
-        smem_k[tid] = (float)k_ptr[tid];
-        smem_q[tid] = (float)q_ptr[tid];
+        smem_k[tid] = k_ptr[tid];
+        smem_q[tid] = q_ptr[tid];
     }
     __syncthreads();
 
     if (tid < v_dim) {
-        float v_i = (float)v_ptr[tid];
+        float v_i = v_ptr[tid];
 
         float hk_dot = 0.0f;
         #pragma unroll 4
